생산자 삭제 시, 생산자 상세도 삭제하도록 수정

+++ src/main/java/com/at/apcss/am/cmns/service/impl/PrdcrServiceImpl.java
@@ -4,6 +4,8 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 
+import javax.annotation.Resource;
+
 import org.egovframe.rte.fdl.cmmn.exception.EgovBizException;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,7 +13,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 
 import com.at.apcss.am.cmns.mapper.PrdcrMapper;
+import com.at.apcss.am.cmns.service.PrdcrDtlService;
 import com.at.apcss.am.cmns.service.PrdcrService;
+import com.at.apcss.am.cmns.vo.PrdcrDtlVO;
 import com.at.apcss.am.cmns.vo.PrdcrVO;
 import com.at.apcss.co.constants.ComConstants;
 import com.at.apcss.co.sys.service.impl.BaseServiceImpl;
@@ -37,6 +41,9 @@ public class PrdcrServiceImpl extends BaseServiceImpl implements PrdcrService {
 
 	@Autowired
 	private PrdcrMapper prdcrMapper;
+
+	@Resource(name="prdcrDtlService")
+	private PrdcrDtlService prdcrDtlService;
 
 	@Override
 	public PrdcrVO selectPrdcr(PrdcrVO prdcrVO) throws Exception {
@@ -87,6 +94,15 @@ public class PrdcrServiceImpl extends BaseServiceImpl implements PrdcrService {
 	@Override
 	public HashMap<String, Object> deletePrdcr(PrdcrVO prdcrVO) throws Exception {
 
+		PrdcrDtlVO prdcrDtlVO = new PrdcrDtlVO();
+		prdcrDtlVO.setApcCd(prdcrVO.getApcCd());
+		prdcrDtlVO.setPrdcrCd(prdcrVO.getPrdcrCd());
+		
+		HashMap<String, Object> rtnObj = prdcrDtlService.deletePrdcrDtl(prdcrDtlVO);
+		if (rtnObj != null) {
+			throw new EgovBizException(getMessageForMap(rtnObj));
+		}
+		
 		if(0 == prdcrMapper.deletePrdcr(prdcrVO)) {
 			throw new EgovBizException(getMessageForMap(ComUtil.getResultMap(ComConstants.MSGCD_ERR_CUSTOM, "삭제 중 오류가 발생 했습니다."))); // E0000	{0}
 		}
